Revert 2.2.29 to 2.2.27
2.2.29 will only build if `--with-brewed-apr` is included. Will push
back to 2.2.29 once solution is found.

Note that if you have full Xcode, instead of only the command line tools,
you will been to add `--with-brewed-zlib` and a fix for that should also
be forthcoming.

With command line tools only, the following builds pass with 2.2.27:
- brew install -v httpd22
- brew install -v httpd22 --with-brewed-openssl
- brew install -v httpd22 --with-brewed-zlib
- brew install -v httpd22 --with-brewed-openssl --with-brewed-zlib
- brew install -v httpd22 --with-brewed-apr
- brew install -v httpd22 --with-brewed-apr --with-brewed-openssl
- brew install -v httpd22 --with-brewed-apr --with-brewed-zlib

With 2.2.29, only the latter three were successful.

With full Xcode, the following builds pass with 2.2.27:
- brew install -v httpd22 --with-brewed-zlib
- brew install -v httpd22 --with-brewed-openssl --with-brewed-zlib
- brew install -v httpd22 --with-brewed-apr --with-brewed-zlib
- brew install -v httpd22 --with-brewed-apr --with-brewed-openssl --with-brewed-zlib

Tests were completed in VMware on fresh systems with nothing else installed.

Closes #41
Closes #40