apr: migrating to Homebrew/Homebrew
Migrating the apr formulae to Homebrew/Homebrew to allow us to deal
with the needs-CLT issue. The CLT is a big package, and if people
don't need everything in it, asking them to install it isn't
necessarily the best user experience we could deliver.

See https://github.com/Homebrew/homebrew/pull/34987

Closes #57